+++ lib/additionals.rb
@@ -3,7 +3,7 @@
 require 'redmine_plugin_kit'
 
 module Additionals
-  VERSION = '4.0.0'
+  VERSION = '4.1.0-main'
 
   MAX_CUSTOM_MENU_ITEMS = 5
   DEFAULT_MODAL_WIDTH = '350px'